Update ReactRoleService.cs
Removed a bunch of fixed values and added them to the json for importing.

Improved dictionary lookup method, and expanded the service to be able to manage roles on a combination of messages instead of just one. Each message can have their own roles so that one message doesn't have 20 different reactions.