Change how the rank embed is displayed
This changes the rank embed to have a field per user rather than 3 inline fields acting as columns.
On mobile, the columns were wrapping and being displayed vertically, making the embed hard to read.

The issue of "Invalid User" mentions is also fixed by having the bot retrieve the username.
(This was actually required since user mentions don't work at all when they are in the field name.)