Send Embed command
!send embed url channel
Send the embed message to the channel provided after waiting for confirmation.

!send embed url channel msgId
Sends embed message to the channel provided, it will check if messageId exists and edit it if possible. If it can't (Not owner of message) it will abort the embed. Will wait for confirmation from user before posting to channel.

- IsValidHost
- GetDownUrlFromUri
2 small helper methods just to reduce duplicated code, few other things that could be improved to reduce code duplication, but eh.