Refactor to not repeat code
All methods now move through SendEmbedToChannel which will act in 1 of 2 ways.

When channel doesn't match context.channel it will ask for confirmation to send the embed first.

If channel is the same as context, the embed will be sent with no confirmation. ie; ``!embed`` with attachment will just post the processed embed into the active channel. where ``!send embed #welcome`` will post the embed in the active channel, and ask for an emote reaction to confirm before sending the embed to the target channel.

If a message id is passed in, then it will attempt to edit the message if it was the author or tell the user it can't edit and abort the embed.