Fix large broken embeds
``displaystyle`` is in all broken embeds due to additional formatting involved. To get around this, we check the extract for ``displaystyle`` if it is found, we do an excessive sanitize of white space and new lines.

This will make results a block of text and destroy natural paragraphs if any exist. But only for results where this is ran.