Small Welcome Service inside the UserService
When a user joins they're still added to a cooldown for welcome, but if they manage to send a message before the 'cooldown' has ended, the bot will welcome them immediately.

This should hopefully reduce the times where someone will have spoken a few lines in offtopic, and then they get the awkward welcome after people are already interacting with them.