Welcome Service improvements
When a user begins to type the bot checks if they're new and haven't been welcomes yet and tries to welcome them before they send a message. This will be a bit more natural then welcoming them immediately after they send a message which seems to confuse a few people.

I've also added a small moderator command to check how many people are currently waiting for welcome and when the next is due. Useless command, but I've seen a couple welcomes happen for people who've joined a week earlier, which shouldn't be the case, just here for debug, only usable by mods/admins.