Simple module for slash commands with a few implemented.
- /roles
can be used to add/remove user roles similar to ``!role add`` this solution uses buttons instead that are hidden and can be used in any channel.

- /help
Shows the help embed in which ever channel used in, only visible to the user. Has button to move between pages. Currently regardless of if there is more than 1 page or not.

- /welcome & /ping
both do the same as their message command counterpart.

- /invite
just shows the invite server from the settings config.