Update to use changed call
- Checks if bot, ignores if it is

Since we no longer have guild user details, we try pull them out of the guild, we check if this is null or not since they're not in the server any more and might already not exist.

If they don't we just post that they left with no time mentioned.